RAG-overview/requirements.txt
sentence-transformers
chromadb
streamlit
google-generativeai
